Rewriting structures to save memory, implemented bitwise ops

+++ src/vm/clone_process.c
@@ -6,7 +6,7 @@
 /*   By: aamadori <aamadori@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/17 19:13:31 by alex              #+#    #+#             */
-/*   Updated: 2019/03/31 16:55:39 by aamadori         ###   ########.fr       */
+/*   Updated: 2019/04/01 17:12:47 by aamadori         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void	vm_init_process(t_vm_state *state, int player_id, size_t address)
 
 	ft_bzero(&new_process, sizeof(t_process));
 	new_process.registers[0].content = byte_order_swap(
-		(t_bigend_buffer){(int64_t)player_id});
+		(t_bigend_buffer){(int32_t)player_id});
 	search = 0;
 	while (search < state->players.length)
 	{
